+++ Controller/AuthController.php
@@ -14,7 +14,7 @@ namespace Zumo\ZumokitBundle\Controller;
 
 use Exception;
 use Zumo\ZumokitBundle\Model\ZumoApp;
-use Zumo\ZumokitBundle\Service\Client\SapiClient;
+use Zumo\ZumokitBundle\Service\Client\ZumokitApiClient;
 use Zumo\ZumokitBundle\Security\Token\JWTEncoder;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Zumo\ZumokitBundle\Service\Request\Validator\RequestValidator;
@@ -42,17 +42,22 @@ class AuthController extends AbstractController
     protected $logger;
 
     /**
-     * @var \Zumo\ZumokitBundle\Model\ZumoApp
+     * @var ZumoApp
      */
     private $app;
 
     /**
-     * @var \Zumo\ZumokitBundle\Service\Client\SapiClient
+     * @var SapiClient
      */
     private $sapi;
 
     /**
-     * @var \Zumo\ZumokitBundle\Security\Token\JWTEncoder
+     * @var ZumokitApiClient
+     */
+    private $zumokitApiClient;
+
+    /**
+     * @var JWTEncoder
      */
     private $tokenEncoder;
 
@@ -69,23 +74,23 @@ class AuthController extends AbstractController
     /**
      * AuthController constructor.
      *
-     * @param \Zumo\ZumokitBundle\Model\ZumoApp                 $app
-     * @param \Zumo\ZumokitBundle\Service\Client\SapiClient     $sapi
-     * @param \Zumo\ZumokitBundle\Security\Token\JWTEncoder $encoder
-     * @param RequestValidator                                  $validator
-     * @param \Psr\Log\LoggerInterface                          $logger
-     * @param UserRepository                                    $repository
+     * @param ZumoApp                   $app
+     * @param SapiClient                $sapi
+     * @param JWTEncoder                $encoder
+     * @param RequestValidator          $validator
+     * @param LoggerInterface           $logger
+     * @param UserRepository            $repository
      */
     public function __construct(
         ZumoApp $app,
-        SapiClient $sapi,
+        ZumokitApiClient $zumokitApiClient,
         JWTEncoder $encoder,
         RequestValidator $validator,
         LoggerInterface $logger,
         UserRepository $repository
     ) {
         $this->app = $app;
-        $this->sapi = $sapi;
+        $this->zumokitApiClient = $zumokitApiClient;
         $this->tokenEncoder = $encoder;
         $this->validator = $validator;
         $this->logger = $logger;
@@ -103,56 +108,29 @@ class AuthController extends AbstractController
      * @param UserInterface|null $user
      * @throws Exception if user object is not valid
      * @throws Exception API key is missing
-     * @throws \GuzzleHttp\Exception\GuzzleException
      * @return JsonResponse
      */
-    public function getZumoKitTokenAction(Request $request, UserInterface $user): JsonResponse
+    public function getZumokitToken(Request $request, UserInterface $userProvidedByAuthorizationToken): JsonResponse
     {
-        try {
-            if (
-                // Here we check only if the user entity has a getter for ID property.
-                // Any symfony implementation will have a username and id, so this is
-                // just a precaution.
-                !($user instanceof UserInterface) ||
-                !method_exists($user, 'getId')
-            ) {
-                throw new Exception("Invalid user object received.");
-            }
+        $appUserId = $userProvidedByAuthorizationToken->getId();
 
-            if (!($request->headers->has('api-key'))) {
-                throw new Exception("Missing API KEY.");
-            }
-        } catch (Exception $exception) {
-            $this->logger->critical(sprintf("Unable to process request, error: %s", $exception->getMessage()));
-            return new JsonResponse(['error' => 'Unauthorized', 'message' => 'Exception occured.'], 401);
+        $doctrine = $this->getDoctrine();
+        $userRepository = $doctrine->getRepository(User::class);
+        $user = $userRepository->findOneBy(['id' => $appUserId]);
+
+        // Check if user exist
+        if (empty($user)) {
+            $this->logger->critical(sprintf('User %s not found.', $appUserId));
+            throw new Exception('User not found.');
         }
 
-        try {
-            $client = new \GuzzleHttp\Client(["base_uri" => $this->sapi->getBaseUri()]);
-            $headers = [
-                'account-id' => (string) $user->getId(),
-                'api-key'    => $this->app->getApiKey(),
-            ];
-
-            $request = new \GuzzleHttp\Psr7\Request(
-                "POST",
-                sprintf('%s/sapi/authentication/token', $this->sapi->getBaseUri()),
-                $headers
-            );
-
-            $response = $client->send($request);
-        } catch (\Exception $exception) {
-            $this->logger->critical(sprintf("Unable to process request, error: %s", $exception->getMessage()));
-            return new JsonResponse(['error' => 'Unauthorized'], 401);
-        }
-
-        return new JsonResponse($response->getBody(), 200, [], true);
+        $result = $this->zumokitApiClient->getTokens($user);
+        return new JsonResponse($result, 200, [], true);
     }
 
     /**
-     * @param \Symfony\Component\HttpFoundation\Request $request
-     *
-     * @return \Symfony\Component\HttpFoundation\JsonResponse
+     * @param Request $request
+     * @return JsonResponse
      */
     public function syncWallets(Request $request)
     {